+++ app/Filament/Immobilisation/Resources/Immobilisation/AssetTransfers/AssetTransferResource.php
@@ -15,14 +15,20 @@ use Filament\Resources\Resource;
 use Filament\Schemas\Schema;
 use Filament\Support\Icons\Heroicon;
 use Filament\Tables\Table;
+use UnitEnum;
 
 class AssetTransferResource extends Resource
 {
     protected static ?string $model = AssetTransfer::class;
 
-    protected static string|BackedEnum|null $navigationIcon = Heroicon::OutlinedRectangleStack;
+    protected static string|BackedEnum|null $navigationIcon = Heroicon::ArrowsUpDown;
 
     protected static ?string $recordTitleAttribute = 'id';
+
+    protected static ?string $modelLabel = 'Transfert d\'asset';
+
+    protected static string | UnitEnum | null $navigationGroup = 'Gestion des Actifs';
+    protected static ?int $navigationSort = 4;
 
     public static function form(Schema $schema): Schema
     {